@@ -5,7 +5,7 @@ readme = "README.md"
 license = "MIT OR Apache-2.0"
 version = "0.2.0"
 authors = ["Matthias Hörmann <mhoermann@gmail.com>"]
-edition = "2021"
+edition = "2024"
 repository = "https://github.com/taladar/sl-map-tools"
 keywords = ["secondlife", "map", "route", "usb_notecard"]
 categories = ["api-bindings"]
