@@ -1,19 +1,19 @@
 #![doc = include_str!("../README.md")]
 
+use chumsky::Parser;
 use chumsky::error::Simple;
 use chumsky::prelude::{any, just, none_of, one_of, take_until};
 use chumsky::text::whitespace;
-use chumsky::Parser;
 
 pub mod avatar_messages;
 pub mod system_messages;
 pub mod utils;
 
 /// represents an event commemorated in the Second Life chat log
-///
-/// large variant warning for clippy is overridden since the Box get in the way
-/// of properly pattern matching
-#[allow(clippy::large_enum_variant)]
+#[expect(
+    clippy::large_enum_variant,
+    reason = "boxing gets in the way of properly pattern matching"
+)]
 #[derive(Debug, Clone, PartialEq)]
 pub enum ChatLogEvent {
     /// line about an avatar (or an object doing things indistinguishable from an avatar in the chat log)
@@ -153,15 +153,15 @@ pub fn chat_log_line_parser() -> impl Parser<char, ChatLogLine, Error = Simple<c
         .try_map(
             |(((((year, month), day), hour), minute), second),
              span: std::ops::Range<usize>| {
-                let second = second.unwrap_or("00".to_string());
+                let second = second.unwrap_or_else(|| "00".to_string());
                 let format = time::macros::format_description!(
                     "[year]/[month]/[day] [hour]:[minute]:[second]"
                 );
                 Ok(Some(
                     time::PrimitiveDateTime::parse(
-                        &format!("{}/{}/{} {}:{}:{}", year, month, day, hour, minute, second),
+                        &format!("{year}/{month}/{day} {hour}:{minute}:{second}"),
                         format,
-                    ).map_err(|e| Simple::custom(span, format!("{:?}", e)))?
+                    ).map_err(|e| Simple::custom(span, format!("{e:?}")))?
                 ))
              }
         )
@@ -181,7 +181,7 @@ pub fn chat_log_line_parser() -> impl Parser<char, ChatLogLine, Error = Simple<c
 
 #[cfg(test)]
 mod test {
-    use std::io::{BufRead, BufReader};
+    use std::io::{BufRead as _, BufReader};
 
     use super::*;
 
@@ -196,7 +196,7 @@ mod test {
 
     /// Error enum for the application
     #[derive(thiserror::Error, Debug)]
-    pub enum TestError {
+    pub(crate) enum TestError {
         /// error loading environment
         #[error("error loading environment: {0}")]
         Env(#[from] envy::Error),
@@ -215,7 +215,7 @@ mod test {
     }
 
     /// determine avatar log dir from avatar name
-    pub fn avatar_log_dir(avatar_name: &str) -> Result<std::path::PathBuf, TestError> {
+    pub(crate) fn avatar_log_dir(avatar_name: &str) -> Result<std::path::PathBuf, TestError> {
         let avatar_dir_name = avatar_name.replace(' ', "_").to_lowercase();
         tracing::debug!("Avatar dir name: {}", avatar_dir_name);
 
@@ -241,14 +241,17 @@ mod test {
             let mut last_line: Option<String> = None;
             for line in file.lines() {
                 let line = line.map_err(TestError::ChatLogLineRead)?;
-                if line.starts_with(" ") || line.is_empty() {
-                    if let Some(ll) = last_line {
-                        last_line = Some(format!("{}\n{}", ll, line));
-                        continue;
-                    }
+                if (line.starts_with(' ') || line.is_empty())
+                    && let Some(ll) = last_line
+                {
+                    last_line = Some(format!("{ll}\n{line}"));
+                    continue;
                 }
                 if let Some(ref ll) = last_line {
-                    #[allow(clippy::panic)]
+                    #[expect(
+                        clippy::panic,
+                        reason = "panics are okay - even intentional - in tests"
+                    )]
                     match chat_log_line_parser().parse(ll.clone()) {
                         Err(e) => {
                             tracing::error!("failed to parse line\n{}", ll);
@@ -270,8 +273,8 @@ mod test {
                             } = parsed_line
                             {
                                 tracing::info!("parsed line\n{}\n{:?}", ll, parsed_line);
-                                if message.starts_with("The message sent to") {
-                                    if let Err(e) =
+                                if message.starts_with("The message sent to") &&
+                                    let Err(e) =
                                         system_messages::chat_message_still_being_processed_message_parser()
                                             .parse(message.to_string())
                                     {
@@ -282,52 +285,48 @@ mod test {
                                                 errors: vec![e.to_owned()],
                                             });
                                         }
-                                    }
                                 }
-                                if message.contains("owned by") && message.contains("gave you") {
-                                    if let Err(e) =
+                                if message.contains("owned by")
+                                    && message.contains("gave you")
+                                    && let Err(e) =
                                         system_messages::object_gave_object_message_parser()
                                             .parse(message.to_string())
-                                    {
-                                        for e in e {
-                                            tracing::debug!(
-                                                "{}",
-                                                utils::ChumskyError {
-                                                    description: "owned by gave you".to_string(),
-                                                    source: message.to_owned(),
-                                                    errors: vec![e.to_owned()],
-                                                }
-                                            );
-                                        }
+                                {
+                                    for e in e {
+                                        tracing::debug!(
+                                            "{}",
+                                            utils::ChumskyError {
+                                                description: "owned by gave you".to_string(),
+                                                source: message.to_owned(),
+                                                errors: vec![e.to_owned()],
+                                            }
+                                        );
                                     }
                                 }
                                 if message.contains("An object named")
                                     && message.contains("gave you this folder")
-                                {
-                                    if let Err(e) =
+                                    && let Err(e) =
                                         system_messages::object_gave_folder_message_parser()
                                             .parse(message.to_string())
-                                    {
-                                        for e in e {
-                                            tracing::debug!(
-                                                "{}",
-                                                utils::ChumskyError {
-                                                    description:
-                                                        "An object named ... gave you this folder"
-                                                            .to_string(),
-                                                    source: message.to_owned(),
-                                                    errors: vec![e.to_owned()],
-                                                }
-                                            );
-                                        }
+                                {
+                                    for e in e {
+                                        tracing::debug!(
+                                            "{}",
+                                            utils::ChumskyError {
+                                                description:
+                                                    "An object named ... gave you this folder"
+                                                        .to_string(),
+                                                source: message.to_owned(),
+                                                errors: vec![e.to_owned()],
+                                            }
+                                        );
                                     }
                                 }
                                 if message.starts_with("Can't rez object")
                                     && message.contains(
                                         "because the owner of this land does not allow it",
                                     )
-                                {
-                                    if let Err(e) =
+                                    && let Err(e) =
                                         system_messages::permission_to_rez_object_denied_message_parser()
                                             .parse(message.to_string())
                                     {
@@ -338,96 +337,89 @@ mod test {
                                                 errors: vec![e.to_owned()],
                                             });
                                         }
-                                    }
                                 }
-                                if message.starts_with("Teleport completed from") {
-                                    if let Err(e) =
+                                if message.starts_with("Teleport completed from")
+                                    && let Err(e) =
                                         system_messages::teleport_completed_message_parser()
                                             .parse(message.to_string())
-                                    {
-                                        for e in e {
-                                            tracing::debug!(
-                                                "{}",
-                                                utils::ChumskyError {
-                                                    description: "teleported completed".to_string(),
-                                                    source: message.to_owned(),
-                                                    errors: vec![e.to_owned()],
-                                                }
-                                            );
-                                        }
+                                {
+                                    for e in e {
+                                        tracing::debug!(
+                                            "{}",
+                                            utils::ChumskyError {
+                                                description: "teleported completed".to_string(),
+                                                source: message.to_owned(),
+                                                errors: vec![e.to_owned()],
+                                            }
+                                        );
                                     }
                                 }
-                                if message.starts_with("[")
+                                if message.starts_with('[')
                                     && message.contains("status.secondlifegrid.net")
-                                {
-                                    if let Err(e) =
+                                    && let Err(e) =
                                         system_messages::grid_status_event_message_parser()
                                             .parse(message.to_string())
-                                    {
-                                        for e in e {
-                                            tracing::debug!(
-                                                "{}",
-                                                utils::ChumskyError {
-                                                    description: "grid status event".to_string(),
-                                                    source: message.to_owned(),
-                                                    errors: vec![e.to_owned()],
-                                                }
-                                            );
-                                        }
+                                {
+                                    for e in e {
+                                        tracing::debug!(
+                                            "{}",
+                                            utils::ChumskyError {
+                                                description: "grid status event".to_string(),
+                                                source: message.to_owned(),
+                                                errors: vec![e.to_owned()],
+                                            }
+                                        );
                                     }
                                 }
-                                if message.starts_with("Object ID:") {
-                                    if let Err(e) =
+                                if message.starts_with("Object ID:")
+                                    && let Err(e) =
                                         system_messages::extended_script_info_message_parser()
                                             .parse(message.to_string())
-                                    {
-                                        for e in e {
-                                            tracing::debug!(
-                                                "{}",
-                                                utils::ChumskyError {
-                                                    description: "extended script info".to_string(),
-                                                    source: message.to_owned(),
-                                                    errors: vec![e.to_owned()],
-                                                }
-                                            );
-                                        }
+                                {
+                                    for e in e {
+                                        tracing::debug!(
+                                            "{}",
+                                            utils::ChumskyError {
+                                                description: "extended script info".to_string(),
+                                                source: message.to_owned(),
+                                                errors: vec![e.to_owned()],
+                                            }
+                                        );
                                     }
                                 }
-                                if message.starts_with("Bridge") {
-                                    if let Err(e) = system_messages::bridge_message_parser()
+                                if message.starts_with("Bridge")
+                                    && let Err(e) = system_messages::bridge_message_parser()
                                         .parse(message.to_string())
-                                    {
-                                        for e in e {
-                                            tracing::debug!(
-                                                "{}",
-                                                utils::ChumskyError {
-                                                    description: "bridge message".to_string(),
-                                                    source: message.to_owned(),
-                                                    errors: vec![e.to_owned()],
-                                                }
-                                            );
-                                        }
+                                {
+                                    for e in e {
+                                        tracing::debug!(
+                                            "{}",
+                                            utils::ChumskyError {
+                                                description: "bridge message".to_string(),
+                                                source: message.to_owned(),
+                                                errors: vec![e.to_owned()],
+                                            }
+                                        );
                                     }
                                 }
 
-                                if message.starts_with("You paid") {
-                                    if let Err(e) = system_messages::sent_payment_message_parser()
+                                if message.starts_with("You paid")
+                                    && let Err(e) = system_messages::sent_payment_message_parser()
                                         .parse(message.to_string())
-                                    {
-                                        for e in e {
-                                            tracing::debug!(
-                                                "{}",
-                                                utils::ChumskyError {
-                                                    description: "sent payment".to_string(),
-                                                    source: message.to_owned(),
-                                                    errors: vec![e.to_owned()],
-                                                }
-                                            );
-                                        }
+                                {
+                                    for e in e {
+                                        tracing::debug!(
+                                            "{}",
+                                            utils::ChumskyError {
+                                                description: "sent payment".to_string(),
+                                                source: message.to_owned(),
+                                                errors: vec![e.to_owned()],
+                                            }
+                                        );
                                     }
                                 }
-                                if message.contains("Take Linden dollars") {
-                                    if let Err(e) = system_messages::object_granted_permission_to_take_money_parser()
+                                if message.contains("Take Linden dollars") &&
+                                    let Err(e) = system_messages::object_granted_permission_to_take_money_parser()
                                         .parse(message.to_string())
                                     {
                                         for e in e {
@@ -440,44 +432,41 @@ mod test {
                                                 }
                                             );
                                         }
-                                    }
                                 }
-                                if message.starts_with("You have offered a calling card") {
-                                    if let Err(e) =
+                                if message.starts_with("You have offered a calling card")
+                                    && let Err(e) =
                                         system_messages::offered_calling_card_message_parser()
                                             .parse(message.to_string())
-                                    {
-                                        for e in e {
-                                            tracing::debug!(
-                                                "{}",
-                                                utils::ChumskyError {
-                                                    description: "offered calling card".to_string(),
-                                                    source: message.to_owned(),
-                                                    errors: vec![e.to_owned()],
-                                                }
-                                            );
-                                        }
+                                {
+                                    for e in e {
+                                        tracing::debug!(
+                                            "{}",
+                                            utils::ChumskyError {
+                                                description: "offered calling card".to_string(),
+                                                source: message.to_owned(),
+                                                errors: vec![e.to_owned()],
+                                            }
+                                        );
                                     }
                                 }
-                                if message.starts_with("Draw Distance set") {
-                                    if let Err(e) =
+                                if message.starts_with("Draw Distance set")
+                                    && let Err(e) =
                                         system_messages::draw_distance_set_message_parser()
                                             .parse(message.to_string())
-                                    {
-                                        for e in e {
-                                            tracing::debug!(
-                                                "{}",
-                                                utils::ChumskyError {
-                                                    description: "draw distance set".to_string(),
-                                                    source: message.to_owned(),
-                                                    errors: vec![e.to_owned()],
-                                                }
-                                            );
-                                        }
+                                {
+                                    for e in e {
+                                        tracing::debug!(
+                                            "{}",
+                                            utils::ChumskyError {
+                                                description: "draw distance set".to_string(),
+                                                source: message.to_owned(),
+                                                errors: vec![e.to_owned()],
+                                            }
+                                        );
                                     }
                                 }
-                                if message.starts_with("Your object") {
-                                    if let Err(e) =
+                                if message.starts_with("Your object") &&
+                                    let Err(e) =
                                         system_messages::your_object_has_been_returned_message_parser()
                                             .parse(message.to_string())
                                     {
@@ -491,7 +480,6 @@ mod test {
                                                 }
                                             );
                                         }
-                                    }
                                 }
                             }
                         }
